Fix condition handling in lists
If it's a condition, fallback to the paragraph loop strategy.

Note: It may actually break some templates that are using loops in lists but are sometimes passing empty values (e.g. null or undefined) to hide the list completely. In these cases they should pass an empty array instead, to make sure the list is removed and not just its content.

Fixes #123